// backend/src/types.ts
export enum UserRole {
    SUPER_ADMIN = 'SUPER_ADMIN',
    MANAGING_EDITOR = 'MANAGING_EDITOR',
    SECTION_EDITOR = 'SECTION_EDITOR',
    JOURNALIST = 'JOURNALIST',
    CONTRIBUTOR = 'CONTRIBUTOR'
}

export enum PostStatus {
    DRAFT = 'DRAFT',
    REVIEW = 'REVIEW',
    EDITOR_APPROVAL = 'EDITOR_APPROVAL',
    SCHEDULED = 'SCHEDULED',
    PUBLISHED = 'PUBLISHED',
    ARCHIVED = 'ARCHIVED'
}

export enum PostType {
    ARTICLE = 'ARTICLE',
    VIDEO = 'VIDEO',
    OPINION = 'OPINION',
    INTERVIEW = 'INTERVIEW',
    MULTIMEDIA_STORY = 'MULTIMEDIA_STORY',
    STATIC_PAGE = 'STATIC_PAGE'
}